pyproject.toml
[build-system]
requires = ["hatch>=1.16", "hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "darc-el"
version = "0.1.0"
description = "DARC-EL: LLM-Based Data Extraction for Assessing Reporting Completeness in Electrocatalysis Literature"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Roland Ramp" },
]
requires-python = ">=3.14"
dependencies = [
  "pydantic>=2.0",
  "pyzotero",
  "python-dotenv",
]

[project.scripts]
zotero-download = "main:main"

[tool.hatch.build.targets.wheel]
include = ["src/main.py"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build.targets.sdist]
include = ["src/main.py", "README.md", "LICENSE", "pyproject.toml"]
